Move functions to turbopack-utils
