+++ packages/next/src/server/dev/turbopack-utils.ts
@@ -12,9 +12,9 @@ import type { AppBuildManifest } from '../../build/webpack/plugins/app-build-man
 import type { BuildManifest } from '../get-page-files'
 import type { NextConfigComplete } from '../config-shared'
 import loadJsConfig from '../../build/load-jsconfig'
-import { posix } from 'path'
-import { readFile } from 'fs/promises'
-import type {
+import { join, posix } from 'path'
+import { readFile, writeFile } from 'fs/promises'
+import {
   APP_BUILD_MANIFEST,
   APP_PATHS_MANIFEST,
   BUILD_MANIFEST,
@@ -23,7 +23,20 @@ import type {
   PAGES_MANIFEST,
   SERVER_REFERENCE_MANIFEST,
   REACT_LOADABLE_MANIFEST,
+  MIDDLEWARE_BUILD_MANIFEST,
+  INTERCEPTION_ROUTE_REWRITE_MANIFEST,
+  MIDDLEWARE_REACT_LOADABLE_MANIFEST,
 } from '../../shared/lib/constants'
+import { writeFileAtomic } from '../../lib/fs/write-atomic'
+import { deleteCache } from '../../build/webpack/plugins/nextjs-require-cache-hot-reloader'
+import {
+  normalizeRewritesForBuildManifest,
+  type ClientBuildManifest,
+  srcEmptySsgManifest,
+} from '../../build/webpack/plugins/build-manifest-plugin'
+import type { SetupOpts } from '../lib/router-utils/setup-dev-bundler'
+import { isInterceptionRouteRewrite } from '../../lib/generate-interception-routes-rewrites'
+import type { Route } from '../../build/swc'
 
 export interface InstrumentationDefinition {
   files: string[]
@@ -226,4 +239,354 @@ export async function readPartialManifest<T>(
     name
   )
   return JSON.parse(await readFile(posix.join(manifestPath), 'utf-8')) as T
+}
+
+export type BuildManifests = Map<string, BuildManifest>
+export type AppBuildManifests = Map<string, AppBuildManifest>
+export type PagesManifests = Map<string, PagesManifest>
+export type AppPathsManifests = Map<string, PagesManifest>
+export type MiddlewareManifests = Map<string, TurbopackMiddlewareManifest>
+export type ActionManifests = Map<string, ActionManifest>
+export type FontManifests = Map<string, NextFontManifest>
+export type LoadableManifests = Map<string, LoadableManifest>
+export type CurrentEntrypoints = Map<string, Route>
+
+export async function loadMiddlewareManifest(
+  distDir: string,
+  middlewareManifests: MiddlewareManifests,
+  pageName: string,
+  type: 'pages' | 'app' | 'app-route' | 'middleware' | 'instrumentation'
+): Promise<void> {
+  middlewareManifests.set(
+    pageName,
+    await readPartialManifest(distDir, MIDDLEWARE_MANIFEST, pageName, type)
+  )
+}
+
+export async function loadBuildManifest(
+  distDir: string,
+  buildManifests: BuildManifests,
+  pageName: string,
+  type: 'app' | 'pages' = 'pages'
+): Promise<void> {
+  buildManifests.set(
+    pageName,
+    await readPartialManifest(distDir, BUILD_MANIFEST, pageName, type)
+  )
+}
+
+export async function loadAppBuildManifest(
+  distDir: string,
+  appBuildManifests: AppBuildManifests,
+  pageName: string
+): Promise<void> {
+  appBuildManifests.set(
+    pageName,
+    await readPartialManifest(distDir, APP_BUILD_MANIFEST, pageName, 'app')
+  )
+}
+
+export async function loadPagesManifest(
+  distDir: string,
+  pagesManifests: PagesManifests,
+  pageName: string
+): Promise<void> {
+  pagesManifests.set(
+    pageName,
+    await readPartialManifest(distDir, PAGES_MANIFEST, pageName)
+  )
+}
+
+export async function loadAppPathManifest(
+  distDir: string,
+  appPathsManifests: AppPathsManifests,
+  pageName: string,
+  type: 'app' | 'app-route' = 'app'
+): Promise<void> {
+  appPathsManifests.set(
+    pageName,
+    await readPartialManifest(distDir, APP_PATHS_MANIFEST, pageName, type)
+  )
+}
+
+export async function loadActionManifest(
+  distDir: string,
+  actionManifests: ActionManifests,
+  pageName: string
+): Promise<void> {
+  actionManifests.set(
+    pageName,
+    await readPartialManifest(
+      distDir,
+      `${SERVER_REFERENCE_MANIFEST}.json`,
+      pageName,
+      'app'
+    )
+  )
+}
+
+export async function loadFontManifest(
+  distDir: string,
+  fontManifests: FontManifests,
+  pageName: string,
+  type: 'app' | 'pages' = 'pages'
+): Promise<void> {
+  fontManifests.set(
+    pageName,
+    await readPartialManifest(
+      distDir,
+      `${NEXT_FONT_MANIFEST}.json`,
+      pageName,
+      type
+    )
+  )
+}
+
+export async function loadLoadableManifest(
+  distDir: string,
+  loadableManifests: LoadableManifests,
+  pageName: string,
+  type: 'app' | 'pages' = 'pages'
+): Promise<void> {
+  loadableManifests.set(
+    pageName,
+    await readPartialManifest(distDir, REACT_LOADABLE_MANIFEST, pageName, type)
+  )
+}
+
+async function writeBuildManifest(
+  distDir: string,
+  buildManifests: BuildManifests,
+  currentEntrypoints: CurrentEntrypoints,
+  rewrites: SetupOpts['fsChecker']['rewrites']
+): Promise<void> {
+  const buildManifest = mergeBuildManifests(buildManifests.values())
+  const buildManifestPath = join(distDir, BUILD_MANIFEST)
+  const middlewareBuildManifestPath = join(
+    distDir,
+    'server',
+    `${MIDDLEWARE_BUILD_MANIFEST}.js`
+  )
+  const interceptionRewriteManifestPath = join(
+    distDir,
+    'server',
+    `${INTERCEPTION_ROUTE_REWRITE_MANIFEST}.js`
+  )
+  deleteCache(buildManifestPath)
+  deleteCache(middlewareBuildManifestPath)
+  deleteCache(interceptionRewriteManifestPath)
+  await writeFileAtomic(
+    buildManifestPath,
+    JSON.stringify(buildManifest, null, 2)
+  )
+  await writeFileAtomic(
+    middlewareBuildManifestPath,
+    `self.__BUILD_MANIFEST=${JSON.stringify(buildManifest)};`
+  )
+
+  const interceptionRewrites = JSON.stringify(
+    rewrites.beforeFiles.filter(isInterceptionRouteRewrite)
+  )
+
+  await writeFileAtomic(
+    interceptionRewriteManifestPath,
+    `self.__INTERCEPTION_ROUTE_REWRITE_MANIFEST=${JSON.stringify(
+      interceptionRewrites
+    )};`
+  )
+
+  const content: ClientBuildManifest = {
+    __rewrites: rewrites
+      ? (normalizeRewritesForBuildManifest(rewrites) as any)
+      : { afterFiles: [], beforeFiles: [], fallback: [] },
+    ...Object.fromEntries(
+      [...currentEntrypoints.keys()].map((pathname) => [
+        pathname,
+        `static/chunks/pages${pathname === '/' ? '/index' : pathname}.js`,
+      ])
+    ),
+    sortedPages: [...currentEntrypoints.keys()],
+  }
+  const buildManifestJs = `self.__BUILD_MANIFEST = ${JSON.stringify(
+    content
+  )};self.__BUILD_MANIFEST_CB && self.__BUILD_MANIFEST_CB()`
+  await writeFileAtomic(
+    join(distDir, 'static', 'development', '_buildManifest.js'),
+    buildManifestJs
+  )
+  await writeFileAtomic(
+    join(distDir, 'static', 'development', '_ssgManifest.js'),
+    srcEmptySsgManifest
+  )
+}
+
+async function writeFallbackBuildManifest(
+  distDir: string,
+  buildManifests: BuildManifests
+): Promise<void> {
+  const fallbackBuildManifest = mergeBuildManifests(
+    [buildManifests.get('_app'), buildManifests.get('_error')].filter(
+      Boolean
+    ) as BuildManifest[]
+  )
+  const fallbackBuildManifestPath = join(distDir, `fallback-${BUILD_MANIFEST}`)
+  deleteCache(fallbackBuildManifestPath)
+  await writeFileAtomic(
+    fallbackBuildManifestPath,
+    JSON.stringify(fallbackBuildManifest, null, 2)
+  )
+}
+
+async function writeAppBuildManifest(
+  distDir: string,
+  appBuildManifests: AppBuildManifests
+): Promise<void> {
+  const appBuildManifest = mergeAppBuildManifests(appBuildManifests.values())
+  const appBuildManifestPath = join(distDir, APP_BUILD_MANIFEST)
+  deleteCache(appBuildManifestPath)
+  await writeFileAtomic(
+    appBuildManifestPath,
+    JSON.stringify(appBuildManifest, null, 2)
+  )
+}
+
+async function writePagesManifest(
+  distDir: string,
+  pagesManifests: PagesManifests
+): Promise<void> {
+  const pagesManifest = mergePagesManifests(pagesManifests.values())
+  const pagesManifestPath = join(distDir, 'server', PAGES_MANIFEST)
+  deleteCache(pagesManifestPath)
+  await writeFileAtomic(
+    pagesManifestPath,
+    JSON.stringify(pagesManifest, null, 2)
+  )
+}
+
+async function writeAppPathsManifest(
+  distDir: string,
+  appPathsManifests: AppPathsManifests
+): Promise<void> {
+  const appPathsManifest = mergePagesManifests(appPathsManifests.values())
+  const appPathsManifestPath = join(distDir, 'server', APP_PATHS_MANIFEST)
+  deleteCache(appPathsManifestPath)
+  await writeFileAtomic(
+    appPathsManifestPath,
+    JSON.stringify(appPathsManifest, null, 2)
+  )
+}
+
+async function writeMiddlewareManifest(
+  distDir: string,
+  middlewareManifests: MiddlewareManifests
+): Promise<void> {
+  const middlewareManifest = mergeMiddlewareManifests(
+    middlewareManifests.values()
+  )
+  const middlewareManifestPath = join(distDir, 'server', MIDDLEWARE_MANIFEST)
+  deleteCache(middlewareManifestPath)
+  await writeFileAtomic(
+    middlewareManifestPath,
+    JSON.stringify(middlewareManifest, null, 2)
+  )
+}
+
+async function writeActionManifest(
+  distDir: string,
+  actionManifests: ActionManifests
+): Promise<void> {
+  const actionManifest = await mergeActionManifests(actionManifests.values())
+  const actionManifestJsonPath = join(
+    distDir,
+    'server',
+    `${SERVER_REFERENCE_MANIFEST}.json`
+  )
+  const actionManifestJsPath = join(
+    distDir,
+    'server',
+    `${SERVER_REFERENCE_MANIFEST}.js`
+  )
+  const json = JSON.stringify(actionManifest, null, 2)
+  deleteCache(actionManifestJsonPath)
+  deleteCache(actionManifestJsPath)
+  await writeFile(actionManifestJsonPath, json, 'utf-8')
+  await writeFile(
+    actionManifestJsPath,
+    `self.__RSC_SERVER_MANIFEST=${JSON.stringify(json)}`,
+    'utf-8'
+  )
+}
+
+async function writeFontManifest(
+  distDir: string,
+  fontManifests: FontManifests
+): Promise<void> {
+  const fontManifest = mergeFontManifests(fontManifests.values())
+  const json = JSON.stringify(fontManifest, null, 2)
+
+  const fontManifestJsonPath = join(
+    distDir,
+    'server',
+    `${NEXT_FONT_MANIFEST}.json`
+  )
+  const fontManifestJsPath = join(distDir, 'server', `${NEXT_FONT_MANIFEST}.js`)
+  deleteCache(fontManifestJsonPath)
+  deleteCache(fontManifestJsPath)
+  await writeFileAtomic(fontManifestJsonPath, json)
+  await writeFileAtomic(
+    fontManifestJsPath,
+    `self.__NEXT_FONT_MANIFEST=${JSON.stringify(json)}`
+  )
+}
+
+async function writeLoadableManifest(
+  distDir: string,
+  loadableManifests: LoadableManifests
+): Promise<void> {
+  const loadableManifest = mergeLoadableManifests(loadableManifests.values())
+  const loadableManifestPath = join(distDir, REACT_LOADABLE_MANIFEST)
+  const middlewareloadableManifestPath = join(
+    distDir,
+    'server',
+    `${MIDDLEWARE_REACT_LOADABLE_MANIFEST}.js`
+  )
+
+  const json = JSON.stringify(loadableManifest, null, 2)
+
+  deleteCache(loadableManifestPath)
+  deleteCache(middlewareloadableManifestPath)
+  await writeFileAtomic(loadableManifestPath, json)
+  await writeFileAtomic(
+    middlewareloadableManifestPath,
+    `self.__REACT_LOADABLE_MANIFEST=${JSON.stringify(json)}`
+  )
+}
+
+export async function writeManifests(
+  opts: SetupOpts,
+  distDir: string,
+  buildManifests: BuildManifests,
+  appBuildManifests: AppBuildManifests,
+  pagesManifests: PagesManifests,
+  appPathsManifests: AppPathsManifests,
+  middlewareManifests: MiddlewareManifests,
+  actionManifests: ActionManifests,
+  fontManifests: FontManifests,
+  loadableManifests: LoadableManifests,
+  currentEntrypoints: CurrentEntrypoints
+): Promise<void> {
+  await writeBuildManifest(
+    distDir,
+    buildManifests,
+    currentEntrypoints,
+    opts.fsChecker.rewrites
+  )
+  await writeAppBuildManifest(distDir, appBuildManifests)
+  await writePagesManifest(distDir, pagesManifests)
+  await writeAppPathsManifest(distDir, appPathsManifests)
+  await writeMiddlewareManifest(distDir, middlewareManifests)
+  await writeActionManifest(distDir, actionManifests)
+  await writeFontManifest(distDir, fontManifests)
+  await writeLoadableManifest(distDir, loadableManifests)
+  await writeFallbackBuildManifest(distDir, buildManifests)
 }
